Refactor game class and create turn class
